feature/stringutils
- Interpolate method moved from main service to a dedicated
  utils class.
