+++ projects/smarti18n/src/lib/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './object-utils';
+export * from './string-utils';
